+++ lib/src/cookie_manager.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/services.dart';
 
 ///Manages the cookies used by WebView instances.