+++ lib/src/http_auth_credentials_database.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'types.dart';
 import 'package:flutter/services.dart';